OOPS+ and Orientation change bug fix, volume button start
Implemented the 'oops plus' button, to set a split at the last split
plus the average, unless the it is clicked before the average (in which
case it just uses now, as this is the best estimate it has).
Also fixed a bug that occurs when the orientation changes after the the
last lap - the screen redraw was always using 'now', rather than using
the end time for the race if the race is ended.

Finally, added support for using the volume buttons to start the timer
(was previously possible to do splits with volume, but now works for
start as well).